refactor: extract Release creation methods into a factory
